# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users:
    get:
      tags:
      - user-rest-controller
      summary: Get all users
      operationId: getAllUsers
      parameters:
      - name: page
        in: query
        description: Page number of the list of users
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Found a page of user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
    put:
      tags:
      - user-rest-controller
      summary: Update user
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User_Mostrar'
        required: true
      responses:
        "200":
          description: User updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
        "401":
          description: unauthorized to update the user
    post:
      tags:
      - user-rest-controller
      summary: Register a new user
      operationId: registerNewUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: User couldn't register
        "200":
          description: User registerd successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get the user by ID
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
        "200":
          description: Found the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
    put:
      tags:
      - user-rest-controller
      summary: Delete a user
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "404":
          description: User wasn't found
        "200":
          description: User deleted succesfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /api/users/image:
    post:
      tags:
      - user-rest-controller
      summary: upload user image
      operationId: uploadImage
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: You are not authorized to upload the image
        "200":
          description: uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
      - user-rest-controller
      summary: Delete user image
      operationId: deleteImage
      responses:
        "200":
          description: Imaged deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          description: You are not authorized to delete the image
        "400":
          description: Request cannot be processed
...
  /api/tournaments/{id}/image:
    get:
      tags:
      - tournament-rest-controller
      summary: download tournament image
      operationId: downloadImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: You are not authorized to download the image
        "404":
          description: tournament not found
        "204":
          description: there's no image
        "200":
          description: downloaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
    post:
      tags:
      - tournament-rest-controller
      summary: upload tournament image
      operationId: uploadImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: tournament not found
        "200":
          description: uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        "401":
          description: You are not authorized to upload the image
    delete:
      tags:
      - tournament-rest-controller
      summary: delete tournament image
      operationId: deleteImage_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: couldn't get to the image
        "401":
          description: You are not authorized to delete the image
        "200":
          description: deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tournament'
        "404":
          description: tournament not found
...
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Get the user conected
      operationId: getActiveUser
      responses:
        "200":
          description: Found the connected user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
        "404":
          description: User conected not found
  /api/users/me/tournaments:
    get:
      tags:
      - user-rest-controller
      summary: Get the tournaments where the user is inscribed
      operationId: getUserTournaments
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Tournament not found
        "200":
          description: Found the tournament
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/me/pairs:
    get:
      tags:
      - user-rest-controller
      summary: Get the user pairs
      operationId: getUserPairs
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Pairs founded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
        "404":
          description: Pairs weren't found
  /api/users/me/matches:
    get:
      tags:
      - user-rest-controller
      summary: Get the user's matches
      operationId: getUserMatches
      responses:
        "200":
          description: Match found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Mostrar'
        "404":
          description: Match not found
  /api/users/image/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Download user image
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Image not found
        "204":
          description: Image field is empty
        "200":
          description: Imaged downloaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User_Mostrar:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        realName:
          type: string
        location:
          type: string
        country:
          type: string
        phone:
          type: string
        numWins:
          type: integer
          format: int32
        numLoses:
          type: integer
          format: int32
        numMatchesPlayed:
          type: number
          format: double
        historicalKarma:
          type: array
          items:
            type: integer
            format: int32
        status:
          type: boolean
        encodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        karma:
          type: integer
          format: int32
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        realName:
          type: string
        location:
          type: string
        country:
          type: string
        phone:
          type: string
        numWins:
          type: integer
          format: int32
        numLoses:
          type: integer
          format: int32
        numMatchesPlayed:
          type: number
          format: double
        historicalKarma:
          type: array
          items:
            type: integer
            format: int32
        status:
          type: boolean
        encodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        karma:
          type: integer
          format: int32
    Tournament:
      type: object
      properties:
        id:
          type: integer
          format: int64
        owner:
          type: string
        tournamentName:
          type: string
        numParticipants:
          type: integer
          format: int32
        numSignedUp:
          type: integer
          format: int32
        rounds:
          type: integer
          format: int32
        about:
          type: string
        ruleset:
          type: string
        location:
          type: string
        inscriptionDate:
          type: string
        startDate:
          type: string
        started:
          type: boolean
...